Stop motors when car is forward for different char
The car wasn't stopping when it was going forward and a command
that doesn't change its direction was submitted. 

A flag representing forward movement was added, and some piece 
of code changed place, so now the car will stop when it's going 
forward and there's an object near it in any condition, i.e, with any
command that doesn't change its direction submitted.